requirements.txt
aiogram == 2.18;
aioredis >= 2.0.1;
asyncpg >= 0.25.0;
loguru >= 0.5.3;
sqlalchemy >= 1.4.31;
pyyaml >= 6.0;